refactor(da/replication): replace network behaviour with new implementation
